feat(upgrade): update

+++ kubeoperator/scripts/upgrade.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 
 PROJECT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
-#source "${PROJECT_DIR}/kubeoperator/scripts/const.sh"
+source "${PROJECT_DIR}/const.sh"
+target=$1
 
 function online_upgrade() {
     export LATEST_KO_VERSION=$(curl -s https://github.com/KubeOperator/KubeOperator/releases/latest/download 2>&1 |grep -Eo "v([0-9]{1,}\.)+[0-9]{1,}")
@@ -11,7 +12,8 @@ function online_upgrade() {
     export KO_INSTALL_URL="https://github.com/KubeOperator/KubeOperator/releases/download/${LATEST_KO_VERSION}/installer-${LATEST_KO_VERSION}.tar.gz"
     export KO_ANSIBLE_URL="https://github.com/KubeOperator/KubeOperator/releases/download/${LATEST_KO_VERSION}/ansible-${LATEST_KO_VERSION}.tar.gz"
     export KO_NEXUS_URL="https://kubeoperator.fit2cloud.com/nexus/nexus-${LATEST_KO_VERSION}.tar.gz"
-     colorMsg $yellow "\n 提示:在线升级请确保当前主机可以正常连接互联网,升级前会先停止KubeOperator服务" | tee -a ${CWD}/upgrade.log
+
+    colorMsg $yellow "\n 提示:在线升级请确保当前主机可以正常连接互联网,升级前会先停止KubeOperator服务" | tee -a ${CWD}/upgrade.log
     if read -p  "升级操作将会升级到目标版本: -$LATEST_KO_VERSION,是否继续 [y/n]: " yn;then
        if [ "$yn" == "Y" ] || [ "$yn" == "y" ] || [ "$target" == "-y" ];then
           colorMsg $green "... 停止KubeOperator服务" | tee -a ${CWD}/upgrade.log
@@ -25,8 +27,7 @@ function online_upgrade() {
           wget --no-check-certificate "${KO_ANSIBLE_URL}" -P $dir_name| tee -a ${CWD}/upgrade.log
           wget --no-check-certificate "${KO_NEXUS_URL}" -P $dir_name| tee -a ${CWD}/upgrade.log
           # untar
-          tar zxvf $dir_name/installer-${LATEST_KO_VERSION}.tar.gz -C $dir_name| tee -a ${CWD}/upgrade.log
-          \cp -rf $dir_name/installer/koctl /usr/local/bin | tee -a ${CWD}/upgrade.log
+          tar zxvf $dir_name/installer-${LATEST_KO_VERSION}.tar.gz -C $dir_name > /dev/null 2>&1| tee -a ${CWD}/upgrade.log
           # 创建 grafana 持久化目录
           if [[ ! -d $KO_BASE/kubeoperator/data/grafana ]];then
             mkdir -p $KO_BASE/kubeoperator/data/grafana
@@ -45,10 +46,11 @@ function online_upgrade() {
           fi
           \cp -rf $dir_name/installer/kubeoperator/conf/* $KO_BASE/kubeoperator/conf/ | tee -a ${CWD}/upgrade.log
           \cp -rf $dir_name/installer/kubeoperator/docker-compose.yml $KO_BASE/kubeoperator/ | tee -a ${CWD}/upgrade.log
+          \cp -rf $dir_name/installer/koctl $KO_BASE/kubeoperator/ | tee -a ${CWD}/upgrade.log
           colorMsg $yellow " ... 拉取镜像" | tee -a ${CWD}/upgrade.log
-          ${CWD}/koctl pull | tee -a ${CWD}/upgrade.log
-          ${CWD}/koctl start | tee -a ${CWD}/upgrade.log
-          ${CWD}/koctl status | tee -a ${CWD}/upgrade.log
+          ${KO_BASE}/kubeoperator/koctl pull | tee -a ${CWD}/upgrade.log
+          ${KO_BASE}/kubeoperator/koctl start | tee -a ${CWD}/upgrade.log
+          ${KO_BASE}/kubeoperator/koctl status | tee -a ${CWD}/upgrade.log
           if [ $(docker ps -a|grep kubeoperator|wc -l) -gt 0 ] && [ $(docker ps -a|grep kubeoperator |egrep "Exit|unhealthy"|wc -l) -eq 0 ];then
             colorMsg $green "升级完成，当前版本: $LATEST_KO_VERSION" | tee -a ${CWD}/upgrade.log
           else
@@ -99,12 +101,9 @@ function upgrade_init() {
     if [[ ! -d ${KO_BASE}/kubeoperator/scripts ]];then
       mkdir -p ${KO_BASE}/kubeoperator/scripts
     fi
+  if [[ ! -f ${KO_BASE}/kubeoperator/scripts/const.sh ]];then
     \cp -rp ${PROJECT_DIR}/* ${KO_BASE}/kubeoperator/scripts/
-    source ${PROJECT_DIR}/const.sh
-#    if [ -f $(pwd)/kubeoperator/scripts/const.sh ];then
-#      source "$(pwd)/kubeoperator/scripts/const.sh"
-#    else
-#    fi
+  fi
 }
 
 function main() {
